+++ src/common/hooks/index.js
@@ -1,2 +1,3 @@
 export * from './useExportConfig';
 export * from './useExportJobScheduler';
+export * from './useSecureDownload';